requirements.txt
tensorflow==2.1.0
absl-py>=0.2.2
astor==0.7.1
bleach>=3.1.4
bokeh==0.12.16
checksumdir==1.0.5
clang
cycler==0.10.0
cxxfilt
decorator==4.4.0
gast==0.2.2
grpcio==1.20.0
h5py==2.9.0
html5lib==0.9999999
humanize==0.5.1
Jinja2>=2.10.1
Keras>=2.2.0
Keras-Applications>=1.0.2
Keras-Preprocessing>=1.0.1
kiwisolver==1.0.1
labm8==0.1.2
Markdown==3.1
MarkupSafe==1.1.1
matplotlib==3.0.3
networkx==2.1
numpy==1.16.3
packaging==19.0
pandas==0.24.2
protobuf==3.7.1
pyparsing==2.4.0
python-dateutil==2.8.0
pytz==2019.1
PyYAML==5.1
scikit-learn==0.20.3
scipy==1.1.0
Send2Trash==1.5.0
six==1.12.0
sklearn==0.0
termcolor==1.1.0
tornado==6.0.2
umap==0.1.1
werkzeug>=0.15.3
wget==3.2